// src/types/colors.ts
type colors =
  | ''
  | 'Aqua'
  | 'Beige'
  | 'Black'
  | 'Blue'
  | 'Brown'
  | 'Colorful'
  | 'Gray'
  | 'Green'
  | 'Orange'
  | 'Pink'
  | 'Purple'
  | 'Red'
  | 'White'
  | 'Yellow';

type ccolor =
  | ''
  | 'Aqua'
  | 'Beige'
  | 'Black'
  | 'Blue'
  | 'Brown'
  | 'Gray'
  | 'Green'
  | 'Orange'
  | 'Pink'
  | 'Purple'
  | 'Red'
  | 'White'
  | 'Yellow';

export type { ccolor, colors };